Clean code, and upgrade dependencies
