+++ pkg/sidecar/config_test.go
@@ -1,9 +1,8 @@
-package main
+package sidecar
 
 import (
 	"testing"
 
-	"github.com/spiffe/spiffe-helper/pkg/sidecar"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -36,12 +35,12 @@ func TestParseConfig(t *testing.T) {
 func TestValidateConfig(t *testing.T) {
 	for _, tt := range []struct {
 		name        string
-		config      *sidecar.Config
+		config      *Config
 		expectError string
 	}{
 		{
 			name: "no error",
-			config: &sidecar.Config{
+			config: &Config{
 				AgentAddress:       "path",
 				SvidFileName:       "cert.pem",
 				SvidKeyFileName:    "key.pem",
@@ -50,7 +49,7 @@ func TestValidateConfig(t *testing.T) {
 		},
 		{
 			name: "no address",
-			config: &sidecar.Config{
+			config: &Config{
 				SvidFileName:       "cert.pem",
 				SvidKeyFileName:    "key.pem",
 				SvidBundleFileName: "bundle.pem",
@@ -59,7 +58,7 @@ func TestValidateConfig(t *testing.T) {
 		},
 		{
 			name: "no SVID file",
-			config: &sidecar.Config{
+			config: &Config{
 				AgentAddress:       "path",
 				SvidKeyFileName:    "key.pem",
 				SvidBundleFileName: "bundle.pem",
@@ -68,7 +67,7 @@ func TestValidateConfig(t *testing.T) {
 		},
 		{
 			name: "no key file",
-			config: &sidecar.Config{
+			config: &Config{
 				AgentAddress:       "path",
 				SvidFileName:       "cert.pem",
 				SvidBundleFileName: "bundle.pem",
@@ -77,7 +76,7 @@ func TestValidateConfig(t *testing.T) {
 		},
 		{
 			name: "no bundle file",
-			config: &sidecar.Config{
+			config: &Config{
 				AgentAddress:    "path",
 				SvidFileName:    "cert.pem",
 				SvidKeyFileName: "key.pem",
